+++ examples/blog/blog/index.js
@@ -3,7 +3,7 @@
  * Module dependencies.
  */
 
-var express = require('./../../lib/express'),
+var express = require('./../../../lib/express'),
     fs = require('fs');
 
 // Export our app as the module
@@ -18,6 +18,15 @@ var posts = JSON.parse(fs.readFileSync(__dirname + '/posts.json', 'utf8'));
 // Set our default view engine to "ejs"
 app.set('view engine', 'ejs');
 
+app.dynamicHelpers({
+    basepath: function(){
+        // "this" is the app, we can
+        // dynamically provide the "home"
+        // setting to all views
+        return this.set('home');
+    }
+});
+
 app.get('/', function(req, res){
     res.render('index', {
         locals: {
@@ -26,10 +35,11 @@ app.get('/', function(req, res){
     });
 });
 
-app.get('/post/:id', function(req, res, params){
+app.get('/post/:id', function(req, res){
+    var id = req.params.id;
     res.render('post', {
         locals: {
-            post: posts[params.id]
+            post: posts[id]
         }
     });
 });