+++ examples/error/app.js
@@ -3,8 +3,7 @@
  * Module dependencies.
  */
 
-var express = require('./../../lib/express'),
-    connect = require('connect');
+var express = require('./../../lib/express');
 
 var app = express.createServer();
 
@@ -13,7 +12,7 @@ app.get('/', function(req, res){
     throw new Error('something broke!');
 });
 
-app.get('/next', function(req, res, params, next){
+app.get('/next', function(req, res, next){
     // We can also pass exceptions to next()
     next(new Error('oh no!'))
 });
@@ -21,6 +20,6 @@ app.get('/next', function(req, res, params, next){
 // The errorHandler middleware in this case will dump exceptions to stderr
 // as well as show the stack trace in responses, currently handles text/plain,
 // text/html, and application/json responses to aid in development
-app.use('/', connect.errorHandler({ dumpExceptions: true, showStack: true }));
+app.use('/', express.errorHandler({ dumpExceptions: true, showStack: true }));
 
 app.listen(3000);