+++ examples/jade/app.js
@@ -3,8 +3,7 @@
  * Module dependencies.
  */
 
-var express = require('./../../lib/express'),
-    connect = require('connect');
+var express = require('./../../lib/express');
 
 // Path to our public directory
 
@@ -14,21 +13,13 @@ var pub = __dirname + '/public';
 // and then serve with connect's staticProvider
 
 var app = express.createServer(
-    connect.compiler({ src: pub, enable: ['sass'] }),
-    connect.staticProvider(pub)
+    express.compiler({ src: pub, enable: ['sass'] }),
+    express.staticProvider(pub)
 );
 
 // Optional since express defaults to CWD/views
 
 app.set('views', __dirname + '/views');
-
-// Enable auto-reloading of view contents when changed
-// with an interval of 1000 milliseconds. Start the app
-// with $ node examples/jade/app.js
-// then alter some views :)
-
-app.set('reload views', 1000);
-// or app.enable('reload views'); for defaults
 
 // Set our default template engine to "jade"
 // which prevents the need for extensions (although you can still mix and match)