+++ examples/mvc/controllers/app.js
@@ -1,0 +1,9 @@
+
+module.exports = {
+    
+    // /
+    
+    index: function(req, res){
+        res.render();
+    }
+};