+++ test/request.test.js
@@ -14,7 +14,7 @@ module.exports = {
     'test #isXMLHttpRequest': function(assert){
         var app = express.createServer();
         
-        app.get('/isxhr', function(req, res, params){
+        app.get('/isxhr', function(req, res){
             assert.equal(req.xhr, req.isXMLHttpRequest);
             res.send(req.isXMLHttpRequest
                 ? 'yeaaa boy'
@@ -117,6 +117,12 @@ module.exports = {
             connect.session({ store: memoryStore })
         );
 
+        app.flashFormatters = {
+            u: function(val){
+                return String(val).toUpperCase();
+            }
+        };
+
         app.get('/', function(req, res){
             assert.eql([], req.flash('info'));
             assert.eql([], req.flash('error'));
@@ -134,6 +140,13 @@ module.exports = {
             req.flash('info', 'Email _sent_.');
             req.flash('info', '<script>');
             assert.eql(['Email <em>sent</em>.', '&lt;script&gt;'], req.flash('info'));
+            
+            req.flash('info', 'Welcome _%s_ to %s', 'TJ', 'something');
+            assert.eql(['Welcome <em>TJ</em> to something'], req.flash('info'));
+
+            req.flash('info', 'Foo %u', 'bar');
+            assert.eql(['Foo BAR'], req.flash('info'));
+
             res.send('ok');
         });