+++ test/response.test.js
@@ -36,6 +36,10 @@ module.exports = {
         app.get('/buffer', function(req, res){
             res.send(new Buffer('wahoo!'));
         });
+        
+        app.get('/noargs', function(req, res, next){
+            res.send();
+        });
 
         assert.response(app,
             { url: '/html' },
@@ -55,9 +59,15 @@ module.exports = {
         assert.response(app,
             { url: '/buffer' },
             { body: 'wahoo!', headers: { 'Content-Type': 'application/octet-stream' }});
+        assert.response(app,
+            { url: '/noargs' },
+            { status: 204 }, function(res){
+                assert.equal(undefined, res.headers['content-type']);
+                assert.equal(undefined, res.headers['content-length']);
+            });
     },
     
-    'test #contentType': function(assert){
+    'test #contentType()': function(assert){
         var app = express.createServer();
         
         app.get('/html', function(req, res){
@@ -71,7 +81,7 @@ module.exports = {
             { body: '<p>yay</p>', headers: { 'Content-Type': 'text/html; charset=utf-8' }});
     },
     
-    'test #attachment': function(assert){
+    'test #attachment()': function(assert){
         var app = express.createServer();
         
         app.get('/style.css', function(req, res){
@@ -79,8 +89,8 @@ module.exports = {
             res.send('some stylezzz');
         });
 
-        app.get('/*', function(req, res, params){
-            res.attachment(params.splat[0]);
+        app.get('/*', function(req, res){
+            res.attachment(req.params[0]);
             res.send('whatever');
         });
         
@@ -100,9 +110,9 @@ module.exports = {
 
         app2.redirect('google', 'http://google.com');
 
-        app2.redirect('blog', function(req, res, params){
-            return params.id
-                ? '/user/' + params.id + '/blog'
+        app2.redirect('blog', function(req, res){
+            return req.params.id
+                ? '/user/' + req.params.id + '/blog'
                 : null;
         });
         
@@ -135,49 +145,56 @@ module.exports = {
         });
 
         app2.get('/user/:id', function(req, res){
+            res.header('X-Foo', 'bar');
             res.redirect('blog');
         });
         
         assert.response(app,
             { url: '/' },
-            { body: '', status: 301, headers: { Location: 'http://google.com' }});
+            { body: 'Redirecting to http://google.com', status: 301, headers: { Location: 'http://google.com' }});
         assert.response(app,
             { url: '/back' },
-            { body: '', status: 302, headers: { Location: '/' }});
+            { body: 'Redirecting to /', status: 302, headers: { Location: '/', 'Content-Type': 'text/plain' }});
         assert.response(app,
             { url: '/back', headers: { Referer: '/foo' }},
-            { body: '', status: 302, headers: { Location: '/foo' }});
+            { body: 'Redirecting to /foo', status: 302, headers: { Location: '/foo' }});
         assert.response(app,
             { url: '/back', headers: { Referrer: '/foo' }},
-            { body: '', status: 302, headers: { Location: '/foo' }});
+            { body: 'Redirecting to /foo', status: 302, headers: { Location: '/foo' }});
         assert.response(app,
             { url: '/home' },
-            { body: '', status: 302, headers: { Location: '/' }});
+            { body: 'Redirecting to /', status: 302, headers: { Location: '/' }});
 
         assert.response(app2,
             { url: '/' },
-            { body: '', status: 301, headers: { Location: 'http://google.com' }});
+            { body: 'Redirecting to http://google.com', status: 301, headers: { Location: 'http://google.com' }});
         assert.response(app2,
             { url: '/back' },
-            { body: '', status: 302, headers: { Location: '/blog' }});
+            { body: 'Redirecting to /blog', status: 302, headers: { Location: '/blog' }});
         assert.response(app2,
             { url: '/home' },
-            { body: '', status: 302, headers: { Location: '/blog' }});
+            { body: 'Redirecting to /blog', status: 302, headers: { Location: '/blog' }});
         assert.response(app2,
             { url: '/google' },
-            { body: '', headers: { Location: 'http://google.com' }});
+            { body: 'Redirecting to http://google.com', headers: { Location: 'http://google.com' }});
         assert.response(app2,
             { url: '/user/12' },
-            { body: '', headers: { Location: '/user/12/blog' }});
+            { body: 'Redirecting to /user/12/blog', headers: { Location: '/user/12/blog', 'X-Foo': 'bar' }});
     },
     
     'test #sendfile()': function(assert){
         var app = express.createServer();
-        
-        app.get('/*', function(req, res, params){
-            var file = params.splat[0];
-            res.sendfile(__dirname + '/fixtures/' + file);
+
+        app.get('/*', function(req, res, next){
+            var file = req.params[0],
+                filePath = __dirname + '/fixtures/' + file;
+            res.sendfile(filePath, function(err, path){
+                assert.equal(path, filePath);
+                if (err) next(err);
+            });
         });
+        
+        app.use(express.errorHandler());
         
         assert.response(app,
             { url: '/user.json' },
@@ -187,7 +204,7 @@ module.exports = {
             { body: '%p Hello World', status: 200, headers: { 'Content-Type': 'application/octet-stream' }});
         assert.response(app,
             { url: '/doesNotExist' },
-            { body: 'Not Found', status: 404 });
+            { body: 'Internal Server Error', status: 500 });
         assert.response(app,
             { url: '/partials' },
             { body: 'Internal Server Error', status: 500 });
@@ -196,12 +213,15 @@ module.exports = {
     'test #download()': function(assert){
         var app = express.createServer();
         
-        app.get('/json', function(req, res, params, next){
-            res.download(__dirname + '/fixtures/user.json', 'account.json');
+        app.get('/json', function(req, res, next){
+            var filePath = __dirname + '/fixtures/user.json';
+            res.download(filePath, 'account.json', function(err, path){
+                assert.equal(filePath, path);
+            });
         });
 
-        app.get('/:file', function(req, res, params, next){
-            res.download(__dirname + '/fixtures/' + params.file);
+        app.get('/*', function(req, res, next){
+            res.download(__dirname + '/fixtures/' + req.params[0]);
         });
         
         assert.response(app,